requirements.txt
python-telegram-bot>=21.7,<22
yt-dlp>=2025.1.0
python-dotenv>=1.0.1
